+++ MinecraftClient/ChatBots/ReplayCapture.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using MinecraftClient.Protocol;
 
 namespace MinecraftClient.ChatBots
 {
